refactor(acceptance): extract scenario 20-path helpers into sibling file
Move file-local interfaces, constants, fixture builder, polling, and
path-mode proxy/login helpers out of 20-preview-path.ts into a new
20-preview-path.helpers.ts, mirroring the 20-preview.helpers.ts
precedent (warren-65f6). Scenario body and scenario const unchanged.

warren-0324
